requirements.txt
fastmcp
Pillow  